# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MaxxTech AI Agent API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Conversation {
  id: number;
  title: string;
  /** @nullable */
  model?: string | null;
  pinned?: boolean;
  messageCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface ConversationInput {
  /** @minLength 1 */
  title: string;
  model?: string;
}

export interface ConversationUpdate {
  /** @minLength 1 */
  title?: string;
  pinned?: boolean;
  model?: string;
}

export type MessageRole = typeof MessageRole[keyof typeof MessageRole];


export const MessageRole = {
  user: 'user',
  assistant: 'assistant',
  system: 'system',
  tool: 'tool',
} as const;

export interface Message {
  id: number;
  conversationId: number;
  role: MessageRole;
  content: string;
  /** @nullable */
  toolCalls?: string | null;
  createdAt: string;
}

export interface MessageInput {
  /** @minLength 1 */
  content: string;
}

export type ToolCallStatus = typeof ToolCallStatus[keyof typeof ToolCallStatus];


export const ToolCallStatus = {
  running: 'running',
  success: 'success',
  error: 'error',
} as const;

export interface ToolCall {
  id: number;
  /** @nullable */
  messageId?: number | null;
  toolName: string;
  /** @nullable */
  input?: string | null;
  /** @nullable */
  output?: string | null;
  status: ToolCallStatus;
  /** @nullable */
  durationMs?: number | null;
  createdAt: string;
}

export interface AgentResponse {
  message: Message;
  toolCalls: ToolCall[];
  /** @nullable */
  thinking?: string | null;
}

export type CodeRunInputLanguage = typeof CodeRunInputLanguage[keyof typeof CodeRunInputLanguage];


export const CodeRunInputLanguage = {
  javascript: 'javascript',
  typescript: 'typescript',
  python: 'python',
  bash: 'bash',
  shell: 'shell',
} as const;

export interface CodeRunInput {
  code: string;
  language: CodeRunInputLanguage;
}

export interface CodeRunResult {
  stdout: string;
  stderr: string;
  exitCode: number;
  durationMs: number;
}

export interface WebSearchInput {
  /** @minLength 1 */
  query: string;
}

export interface SearchResultItem {
  title: string;
  url: string;
  snippet: string;
}

export interface WebSearchResult {
  query: string;
  results: SearchResultItem[];
}

export interface AgentStats {
  totalConversations: number;
  totalMessages: number;
  totalToolCalls: number;
  modelsUsed: string[];
}

export type SettingsTheme = typeof SettingsTheme[keyof typeof SettingsTheme];


export const SettingsTheme = {
  light: 'light',
  dark: 'dark',
  system: 'system',
} as const;

export interface Settings {
  id: number;
  agentName: string;
  model: string;
  systemPrompt: string;
  enableCodeExecution: boolean;
  enableWebSearch: boolean;
  /** @nullable */
  maxTokens?: number | null;
  /** @nullable */
  temperature?: number | null;
  theme: SettingsTheme;
  createdAt: string;
  updatedAt: string;
}

export type SettingsUpdateTheme = typeof SettingsUpdateTheme[keyof typeof SettingsUpdateTheme];


export const SettingsUpdateTheme = {
  light: 'light',
  dark: 'dark',
  system: 'system',
} as const;

export interface SettingsUpdate {
  agentName?: string;
  model?: string;
  systemPrompt?: string;
  enableCodeExecution?: boolean;
  enableWebSearch?: boolean;
  /** @nullable */
  maxTokens?: number | null;
  /** @nullable */
  temperature?: number | null;
  theme?: SettingsUpdateTheme;
}