# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MaxxTech AI Agent API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /conversations/{id}/send:
    post:
      operationId: sendMessage
      tags: [agent]
      summary: Send a message and get AI response with tool calls
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "200":
          description: AI response with tool usage
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentResponse"
  /agent/run-code:
    post:
      operationId: runCode
      tags: [agent]
      summary: Execute code in a sandbox
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CodeRunInput"
      responses:
        "200":
          description: Code execution result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CodeRunResult"
  /agent/web-search:
    post:
      operationId: webSearch
      tags: [agent]
      summary: Search the web
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebSearchInput"
      responses:
        "200":
          description: Web search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebSearchResult"
...
components:
  schemas:
...
    Message:
      type: object
      required: [id, conversationId, role, content, createdAt]
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
          enum: [user, assistant, system, tool]
        content:
          type: string
        toolCalls:
          type: ["string", "null"]
        createdAt:
          type: string
    MessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
    AgentResponse:
      type: object
      required: [message, toolCalls]
      properties:
        message:
          $ref: "#/components/schemas/Message"
        toolCalls:
          type: array
          items:
            $ref: "#/components/schemas/ToolCall"
        thinking:
          type: ["string", "null"]
    ToolCall:
      type: object
      required: [id, toolName, status, createdAt]
      properties:
        id:
          type: integer
        messageId:
          type: ["integer", "null"]
        toolName:
          type: string
        input:
          type: ["string", "null"]
        output:
          type: ["string", "null"]
        status:
          type: string
          enum: [running, success, error]
        durationMs:
          type: ["integer", "null"]
        createdAt:
          type: string
    CodeRunInput:
      type: object
      required: [code, language]
      properties:
        code:
          type: string
        language:
          type: string
          enum: [javascript, typescript, python, bash, shell]
    CodeRunResult:
      type: object
      required: [stdout, stderr, exitCode, durationMs]
      properties:
        stdout:
          type: string
        stderr:
          type: string
        exitCode:
          type: integer
        durationMs:
          type: integer
    WebSearchInput:
      type: object
      required: [query]
      properties:
        query:
          type: string
          minLength: 1
    WebSearchResult:
      type: object
      required: [query, results]
      properties:
        query:
          type: string
        results:
          type: array
          items:
            $ref: "#/components/schemas/SearchResultItem"
    SearchResultItem:
      type: object
      required: [title, url, snippet]
      properties:
        title:
          type: string
        url:
          type: string
        snippet:
          type: string
...