# lib/api-zod/src/generated/types/agentResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MaxxTech AI Agent API
 * OpenAPI spec version: 0.1.0
 */
import type { Message } from './message';
import type { ToolCall } from './toolCall';

export interface AgentResponse {
  message: Message;
  toolCalls: ToolCall[];
  /** @nullable */
  thinking?: string | null;
}