# lib/api-zod/src/generated/types/searchResultItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MaxxTech AI Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface SearchResultItem {
  title: string;
  url: string;
  snippet: string;
}