# lib/api-zod/src/generated/types/toolCall.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MaxxTech AI Agent API
 * OpenAPI spec version: 0.1.0
 */
import type { ToolCallStatus } from './toolCallStatus';

export interface ToolCall {
  id: number;
  /** @nullable */
  messageId?: number | null;
  toolName: string;
  /** @nullable */
  input?: string | null;
  /** @nullable */
  output?: string | null;
  status: ToolCallStatus;
  /** @nullable */
  durationMs?: number | null;
  createdAt: string;
}