# lib/api-zod/src/generated/types/webSearchInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MaxxTech AI Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface WebSearchInput {
  /** @minLength 1 */
  query: string;
}